MSBuild: Add 64-bit builds for the client & N3Base_client projects
Fix numerous warnings as a result (warnings are treated as errors in these builds).

+++ Client/N3Base/N3AnimControl.h
@@ -131,9 +131,10 @@ public:
 		WriteFile(hFile, &fFrmStrike1, 4, &dwRWC, nullptr);
 
 		// 이름 읽기..
-		nL = szName.size();
+		nL = static_cast<int>(szName.size());
 		WriteFile(hFile, &nL, 4, &dwRWC, nullptr);
-		if(nL > 0) WriteFile(hFile, szName.c_str(), nL, &dwRWC, nullptr);
+		if (nL > 0)
+			WriteFile(hFile, szName.c_str(), nL, &dwRWC, nullptr);
 	}
 
 #ifdef _N3TOOL
